+++ projects/test-app/src/app/state/feature/feature.selector.ts
@@ -1,0 +1,4 @@
+import { createFeatureSelector } from '@ngrx/store';
+import { FeatureState } from './feature.state';
+
+export const featureState = createFeatureSelector<FeatureState>('feature');