+++ projects/test-app/src/app/state/feature/feature.state.ts
@@ -1,0 +1,10 @@
+import { ActionReducerMap } from '@ngrx/store';
+import { customerReducer, ICustomerState } from './customer.state';
+
+export interface FeatureState {
+  customer: ICustomerState;
+}
+
+export const featureReducer: ActionReducerMap<FeatureState> = {
+  customer: customerReducer
+};